Separo las demas acciones en GenerarCertificados
